+++ packages/cli/src/render.tsx
@@ -11,7 +11,6 @@ import fs from 'fs';
 import os from 'os';
 import path from 'path';
 import {Internals} from 'remotion';
-import {deleteDirectory} from './delete-directory';
 import {getCliOptions} from './get-cli-options';
 import {getCompositionId} from './get-composition-id';
 import {initializeRenderCli} from './initialize-render-cli';
@@ -100,10 +99,6 @@ export const render = async () => {
 	const outputDir = shouldOutputImageSequence
 		? absoluteOutputFile
 		: await fs.promises.mkdtemp(path.join(os.tmpdir(), 'react-motion-render'));
-
-	if (!outputDir) {
-		throw new Error('Assertion error: Expected outputDir to not be null');
-	}
 
 	Log.verbose('Output dir', outputDir);
 
@@ -207,7 +202,6 @@ export const render = async () => {
 			updateRenderProgress();
 		},
 		openedBrowser,
-		outputDir,
 		overwrite,
 		parallelism,
 		pixelFormat,
@@ -250,20 +244,7 @@ export const render = async () => {
 	close();
 	Log.verbose('Cleaning up...');
 	try {
-		if (process.platform === 'win32') {
-			// Properly delete directories because Windows doesn't seem to like fs.
-			await deleteDirectory(outputDir);
-			await deleteDirectory(bundled);
-		} else {
-			await Promise.all([
-				(fs.promises.rm ?? fs.promises.rmdir)(outputDir, {
-					recursive: true,
-				}),
-				(fs.promises.rm ?? fs.promises.rmdir)(bundled, {
-					recursive: true,
-				}),
-			]);
-		}
+		await RenderInternals.deleteDirectory(bundled);
 	} catch (err) {
 		Log.warn('Could not clean up directory.');
 		Log.warn(err);
